+++ developer/src/server/src/win32/trayicon/addon-src/TrayWrapper.cpp
@@ -1,0 +1,8 @@
+#include "TrayIcon.h"
+
+Object Init(Env env, Object exports) {
+	CTrayIconContainer::Init(env, exports);
+	return exports;
+}
+
+NODE_API_MODULE(addon, Init)